Update walt.c
Fixed SCHED_PRINT formatting

+++ kernel/sched/walt/walt.c
@@ -206,7 +206,9 @@ early_param("sched_predl", set_sched_predl);
 __read_mostly unsigned int walt_scale_demand_divisor;
 #define scale_demand(d) ((d)/walt_scale_demand_divisor)
 
-#define SCHED_PRINT(arg)        printk_deferred("%s=%llu", #arg, arg)
+// #define SCHED_PRINT(arg)        printk_deferred("%s=%llu", #arg, arg)
+#define SCHED_PRINT(arg) printk_deferred("%s=%llu", #arg, (unsigned long long)(arg))
+
 #define STRG(arg)               #arg
 
 static inline void walt_task_dump(struct task_struct *p)
